Move LfLimit to LfBranch
