+++ src/main/java/com/powsybl/openloadflow/network/impl/AbstractLfBranch.java
@@ -4,10 +4,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openloadflow.network;
+package com.powsybl.openloadflow.network.impl;
 
 import com.powsybl.iidm.network.*;
-import com.powsybl.openloadflow.network.impl.Transformers;
+import com.powsybl.openloadflow.network.*;
 import com.powsybl.openloadflow.util.Evaluable;
 import org.apache.commons.math3.util.FastMath;
 import org.slf4j.Logger;
@@ -19,37 +19,6 @@ import java.util.*;
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
 public abstract class AbstractLfBranch extends AbstractElement implements LfBranch {
-
-    public static final class LfLimit {
-
-        private int acceptableDuration;
-        private final double value;
-
-        private LfLimit(int acceptableDuration, double value) {
-            this.acceptableDuration = acceptableDuration;
-            this.value = value;
-        }
-
-        private static LfLimit createTemporaryLimit(int acceptableDuration, double valuePerUnit) {
-            return new LfLimit(acceptableDuration, valuePerUnit);
-        }
-
-        private static LfLimit createPermanentLimit(double valuePerUnit) {
-            return new LfLimit(Integer.MAX_VALUE, valuePerUnit);
-        }
-
-        public int getAcceptableDuration() {
-            return acceptableDuration;
-        }
-
-        public double getValue() {
-            return value;
-        }
-
-        public void setAcceptableDuration(int acceptableDuration) {
-            this.acceptableDuration = acceptableDuration;
-        }
-    }
 
     private static final Logger LOGGER = LoggerFactory.getLogger(AbstractLfBranch.class);
