+++ src/main/java/com/powsybl/openloadflow/sa/AbstractSecurityAnalysis.java
@@ -161,14 +161,14 @@ public abstract class AbstractSecurityAnalysis {
         }
     }
 
-    protected static LimitViolation createLimitViolation1(LfBranch branch, AbstractLfBranch.LfLimit temporaryLimit1,
+    protected static LimitViolation createLimitViolation1(LfBranch branch, LfBranch.LfLimit temporaryLimit1,
                                                           LimitViolationType type, double scale, double value) {
         return new LimitViolation(branch.getId(), type, null,
                 temporaryLimit1.getAcceptableDuration(), temporaryLimit1.getValue() * scale,
                 (float) 1., value * scale, Branch.Side.ONE);
     }
 
-    protected static LimitViolation createLimitViolation2(LfBranch branch, AbstractLfBranch.LfLimit temporaryLimit2,
+    protected static LimitViolation createLimitViolation2(LfBranch branch, LfBranch.LfLimit temporaryLimit2,
                                                           LimitViolationType type, double scale, double value) {
         return new LimitViolation(branch.getId(), type, null,
                 temporaryLimit2.getAcceptableDuration(), temporaryLimit2.getValue() * scale,
